+++ src/index.js
@@ -1,5 +1,6 @@
 import './app/index.css'
 import './utils/APIErrorHandler'
+import './utils/ConsoleLogger'
 
 import React from 'react'
 import ReactDOM from 'react-dom/client'